requirements.txt
anthropic>=0.40.0
tavily-python>=0.5.0
sendgrid>=6.11.0
python-dotenv>=1.0.0
pytz>=2024.1
requests>=2.31.0
yfinance>=0.2.0
feedparser>=6.0.0
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
